requirements.txt
botcity-framework-core
pyautogui
opencv-python
requests
pygetwindow
numpy
pillow
google-genai
python-dotenv